requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.4
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-timezone-field==7.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
exceptiongroup==1.2.2
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg-pool==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-django==4.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.2
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
